+++ client/pages/create-listing.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import CreateListingFormDetails from './create-listing-form-details';
-import CreateListingFormLocation from './create-listing-form-location';
-import CreateListingPublished from './create-listing-published';
-import Redirect from './redirect';
+import CreateListingFormDetails from '../components/create-listing-form-details';
+import CreateListingFormLocation from '../components/create-listing-form-location';
+import CreateListingPublished from '../components/create-listing-published';
+import Redirect from '../components/redirect';
 
 export default class CreateListingFormParent extends React.Component {
   constructor(props) {
